requirements.txt
streamlit
pandas
pymongo
python-dotenv
google-generativeai
fpdf
requests
